Add new hphp external API hphp_thread_init/hphp_thread_exit
Summary:
For the ThreadLocalNoCheck to work correctly, we need to call init_thread_locals
at least once for each thread before HPHP is ready for use. I added two
new APIs as wrappers for HPHP thread init/exit actions. After that I will
update code to call the new API and remove the call inside
hphp_session_init.

Task ID: #

Blame Rev:

Reviewers:
qigao, mwilliams
CC:
hphp-diffs@lists
Test Plan:
make fast_tests
make slow_tests

Revert Plan:

Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 251684

+++ src/runtime/base/thread_init_fini.cpp
@@ -45,8 +45,8 @@ void init_thread_locals(void *arg /* = NULL */) {
 }
 
 void finish_thread_locals(void *arg /* = NULL */) {
-  g_context.destroy();
-  g_persistentObjects.destroy();
+  if (g_context.getNoCheck()) g_context.destroy();
+  if (g_persistentObjects.getNoCheck()) g_persistentObjects.destroy();
 }
 
 static class SetThreadInitFini {
